chrome-extension/backend/requirements.txt
fastapi
uvicorn
cohere
pydantic
google-generativeai
